imx6: ventana: default GW52xx USB OTG to front-panel
The GW52xx revision-B inverts the OTG select such that an low selects
the PCI socket and a high selects the front-panel. Therefore on powerup
the power-on default of a 100k PU selects front-panel and we keep this setting
in the kernel. If the user wants to route it to the PCIe socket, they need to
set it low.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>