imx6: add ldo-bypass enable
When an external PMIC is used for VDD_SOC and VDD_ARM you can save power by
bypassing the internal LDO's provided by the anantop regulator as long as
you are running less than 1.2GHz. Additionally, the CPU operating setpoints
that define the relationship between CPU frequency and VDD_ARM / VDD_SOC
voltages define the voltages assuming the LDO is bypassed. If the LDO is not
bypassed, the external PMIC regulator voltages need to be increased by
125mV for correct regulator supply voltage.

Freescale added the ldo-bypass property to the gpc bindings, but only used it
in their bootloader. This patch moves the use of it to the kernel where it
should be.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>