ASoC: tda1997x: constrain sample-rate
commit 0b53f6c8af2dbd49b05f12f751b5fb86dae7f859 added validation of the
sample-rate requested in the hw_params callback. This however does not play
well with ALSA applications as they may fail setting a sample-rate which the
driver clearly stated was ok.

Instead, the better approach is to use the startup callback to constrain
the sample-rate to the only value currently allowable (the source HDMI rate).
With this, the hw_params validation can be removed as it will never fail.

It is important to note that an ALSA application can ask for example a 44.1kHz
rate on a 48kHz stream and the call will not fail. It is the responsibility
of the ALSA application to check what rate is actually being used.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>