tda1997x: fix audio output format details
The tda1997x has an audio output formatter that re-formats the HDMI source
audio depending on the layout (connection to the SoC). Therefore, the layout
dictates the output channels and sample-size, whereas the HDMI source
dictates the output sample-rate (because it is not changed). We also add
tracking the source channels which is obtained from the details of an
Audio InfoFrame.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>