tda1997x: fixed EDID handling
The TDA1997x core supports multiple inputs which share an EDID structure
with the exception of the Source Physical Address (SPA) used for CEC address
and the affected checksum stored in the CEA extension block.

This patch simplifies the configuration of EDID by not requiring a hard-coded
offset (which was set correcty for the EDID and thus stomping on the data).
It does this by parsing the CEA extension block and finding the SPA offset.
Additionally, the SPA for InputA is passed into the EDID and InputB is given
that + 1.

This resolves several issues caused by a corrupt EDID.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>