ventana: use pendown handler from tsc2007
As our tsc2007 has an interrupt back to the CPU it makes more sense to
provide a pendown handler that uses the state of this pin to determine
if the pen is up or down instead of trying to caculate it using x_plat_ohms
and dealing with preasure sensitivity.

This resolves sensitivity issues with the tsc2007 on the GW16109 display board.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>