Revert "ENGR00162649 sgtl5000: reduce the DAC volume"
This reverts commit c7b2146b0b90f8cdeef2b6a012bff898517c4a96.

Clamping the DAC volume as low as this produces a max peak-to-peak Line-out
level of 340mV which is too small. I think Freescale was perhaps working around
some noisy hardware with this hack. The line-out quality of the sgtl5000 on
the Ventana boards sounds fine with 100% volume at the previous levels

Signed-off-by: Tim Harvey <tharvey@gateworks.com>