pwm: pca9685: do not clobber duty cycle on enable/disable
The pca9685 enable/disable functions overwrite the on/off registers which ends
up clobbering their current values. This causes an issue when used as a pwm-led
as that code configures the pwm before enabling/disabling it (as it should).

Instead, use the fact that the FULL_OFF bit takes precendence over all others
and only touch it during enable/disable. We also need to make sure to clear
the FULL_ON bit when configuring the pwm if it is not fully on.

Cc: Axel Lin <axel.lin@ingics.com>
Cc: Steffen Trumtrar <s.trumtrar@pengutronix.de>
Cc: Thierry Reding <thierry.reding@gmail.com>

Signed-off-by: Tim Harvey <tharvey@gateworks.com>