ventana: configure screen resoultion and inverted axis for ft5x06
The DLC700MGT4 that the ft5x06 is on, is a 1024x600 resolution display
which does not agree with the values read from the ft5x06 registers and
reported to userspace. Additionally its 0,0 value is at the bottom right
of the display instead of the top left as it is bonded.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>