ventana: switch to ldo-enabled mode
The IMX6 has internal LDO's for VDD_ARM and VDD_SOC which can be used with a
fixed voltage PMIC. However, if you have a PMIC capable of varying its output
voltages, you can bypass the IMX6 LDO's in what is known as 'ldo-bypass' mode
in which case the IMX6 internal LDO's are bypassed (their FET's are set to
their maximum voltage effectively opening them completely) and the PMIC
adjusts the voltage rails per the imx6-cpufreq driver which implements
Dynamic Voltage and Frequency Scaling (DVFS).

The advantage of ldo-bypass mode is that it improves CPU power efficiency
by using the more efficient switching regulators in the PMIC which reduces
overall power consumption as well as the CPU thermal output power. The
downside of ldo-bypass mode is that switching regulators are inherently
more noisy that LDO's and in certain cases according to Freescale such as
when running at 1.2GHz for processor grades that support this frequency
ldo-enabled mode is required.

We have always defaulted to ldo-bypass mode however we have found that the
voltage rails and are much more stable operating in ldo-enabled mode.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>