defconfig: enable DSA support for MV88E6xxx (GW16083)
This enables DSA support for the MV88E6176 7-port switch on the GW16083
Ethernet Expansion Mezzanine. To register the switch with DSA the CPU
interface to the GW16083 needs to be brought up. This will typically be the
2nd NIC (eth1) registered.

DSA support treats each physical port as its own network interface. This will
add 'lan1' through 'lan6' ethernet devices representing the 6 external ports
on the GW16083.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>