defconfig: add various LED trigger drivers
Note that the user1 LED (front-panel green on most boards) is configured
in the device-tree to use the 'heartbeat' trigger, yet that trigger was not
built into the kernel. Now that it is, expect to see that LED heartbeating
on its own. If you want to change this you can do so in userspace by setting
/sys/class/leds/user1/trigger for example:

echo none > /sys/class/leds/user1/trigger

Signed-off-by: Tim Harvey <tharvey@gateworks.com>