Revert "gsc: add pm_power_off hook"
This reverts commit c382364257bf50a8d373ee72392da9855b071014.

Using the GSC to disable the board's primary power supply means the board
can only then be powered on again by pressing the user pushbutton (or by
removing the GSC battery for 10+ seconds to 'reset' the GSC firmware).

This turns out to be a rather horrible choice for boards that have
sub-loaded the user pushbutton or for boards in an enclosure/system where
the button is not accessible. While we could add a kernel command-line
argument to control this we will simply revert the feature for now.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>