ARM: dts: imx6qdl: switch esdhc3 and esdhc2 order so emmc enumerates first
This is a bit of a hack. Until we find a more elegant way to force mmcblk
instance assignments reversing the order of the esdhc's in the device-tree
will cause them to enumerate in the order we wish. We want EMMC to always be
/dev/mmcblk0 (as its always present on boards that have it) and microSD to
always be /dev/mmcblk1 (as a device may not be inserted).

On all Ventana boards that have EMMC + microSD, EMMC is sdhc2 and microSD is
sdhc3.

Signed-off-by: Robert Jones <rjones@gateworks.com>