mfd: tda1997x: disable HS/VS outputs for BT656
In BT656 mode HS/VS are not used and sync indications are embedded in the
data instead. Disabling the HS/VS pin output has no effect in this mode
so we will disable them as a workaround for an errata on the GW5510-C where
the HS signal is connected improperly.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>