Added missing kernel configs for bluetooth LE GATT
Prior to this patch bluetooth devices were able to be seen through
the system settings and paired with but bluetooth low energy devices
communicating via advertisements on the GATT profile could not be
seen or connected. Inspecting external/bluetooth/bluez/android/README
showed a list of necessary kernel configs, some of which were not
included in the default kernel configuration.

Signed-off-by: Robert Jones <rjones@gateworks.com>