Revert "gwventana_bootscript: enable ubi fastmap autoconvert"
This reverts commit 20b52ce25d631cd47e2be4ae010a49d2643eeecd.

UBI Fastmap is on optional feature which stores the physical to
logical eraseblock relations in a checkpoint (called fastmap) to
reduce the initialization time of UBI. Without UBI Fastmap, the
kernel init time of UBI is proportional to the number of physical
erase blocks on the FLASH device and with Fastmap enabled the scan
time is limited to a fixed number of blocks.

While enabling this saves about 5 seconds in Linux init the
downside is that you wear out the first 64blocks of your NAND faster.
Therefore we will leave it up to the user to decide if they want to use
this if attach/boot time is more critical than NAND liftetime.